Rename ValuesController to DummyController
